Acomodo código para test e intento 1 test
